Moved important functions to new helpres folder
