+++ bot/bot.py
@@ -1,10 +1,6 @@
-import datetime
-from collections import OrderedDict
-
 from aiohttp import ClientSession
 import asyncio
 from discord.ext import commands, tasks
-import discord
 from mysql.connector import MySQLConnection
 from typing import Any
 
@@ -23,7 +19,6 @@ from bot.cogs.ping_command import PingCommand
 from bot.cogs.award_command import AwardCommand
 
 from bot.views.claim_view import ClaimView
-from bot.views.affirm_view import AffirmView
 from bot.views.check_view import CheckView
 from bot.views.check_view_red import CheckViewRed
 from bot.views.resolve_ping_view import ResolvePingView
@@ -35,17 +30,11 @@ from bot.views.force_complete_view import ForceCompleteView
 from bot.views.force_unclaim_view import ForceUnclaimView
 
 from bot.models.outage import Outage
-from bot.models.checked_claim import CheckedClaim
-from bot.models.user import User
 from bot.models.team import Team
-from bot.models.pending_ping import PendingPing
-from bot.models.ping import Ping
-from bot.models.team_point import TeamPoint
 
-from bot.status import Status
-
-from bot.helpers import LeaderboardResults
-from bot.helpers import is_working_time
+from bot.helpers.other import *
+from bot.helpers.leaderboard_helpers import *
+from bot.helpers.pings_and_kudos import *
 
 
 class Bot(commands.Bot):
@@ -186,101 +175,11 @@ class Bot(commands.Bot):
             case_channel = await self.fetch_channel(self.cases_channel)
             for pp in pending_pings:
                 if pp.severity.lower() != "kudos":
-                    await self._send_ping(pp, case_channel)
+                    await send_pending_ping(self, pp, case_channel)
                 else:
-                    await self._send_kudos(pp, case_channel)
+                    await send_pending_kudos(pp, case_channel)
 
                 await asyncio.sleep(5)  # prevent rate limiting
-
-    async def _send_ping(self, pp: PendingPing, case_channel: discord.TextChannel):
-        now = datetime.datetime.now()
-        # Ping the case as normal
-        claim = CheckedClaim.from_checker_message_id(self.connection, pp.checker_message_id)
-        claim.change_status(self.connection, Status.PINGED)
-
-        tech = await self.fetch_user(claim.tech.discord_id)
-        lead = await self.fetch_user(claim.lead.discord_id)
-
-        # During working time, send ping as normal
-        fb_embed = discord.Embed(colour=discord.Color.red(), timestamp=now)
-
-        fb_embed.description = f"<@{tech.id}>, this case has been pinged by <@{lead.id}>."
-
-        fb_embed.add_field(name="Reason", value=str(pp.description), inline=False)
-
-        # Add a to-do message if none is passed in
-        if len(str(pp.to_do)) == 0:
-            fb_embed.add_field(name="To Do",
-                               value="Review and let us know if you have any questions!",
-                               inline=False)
-        else:
-            fb_embed.add_field(name="To Do", value=str(pp.to_do), inline=False)
-
-        fb_embed.add_field(name="", value=str("*Note: Please review this information and take actions during work hours, not after!*"))
-        fb_embed.set_author(name=f"{claim.case_num}", icon_url=f'{tech.display_avatar}')
-        fb_embed.set_footer(text=f"{pp.severity} severity level")
-
-        # Create thread
-        thread = await case_channel.create_thread(
-            name=f"{claim.case_num}",
-            message=None,
-            auto_archive_duration=4320,
-            type=discord.ChannelType.private_thread,
-            reason="Case has been pinged.",
-            invitable=False
-        )
-
-        # Add users to thread and send message
-        await thread.add_user(tech)
-        await thread.add_user(lead)
-        message = await thread.send(embed=fb_embed, view=AffirmView(self))
-
-        # Create ping object
-        ping = Ping(thread.id, message.id, str(pp.severity), str(pp.description))
-        ping.add_to_database(self.connection)
-        claim.add_ping_thread(self.connection, thread.id)
-
-        # Remove pending ping
-        pp.remove_from_database(self.connection)
-
-    async def _send_kudos(self, pp: PendingPing, case_channel: discord.TextChannel):
-        now = datetime.datetime.now()
-        claim = CheckedClaim.from_checker_message_id(self.connection, pp.checker_message_id)
-        tech = await self.fetch_user(claim.tech.discord_id)
-
-        fb_embed = discord.Embed(
-            description=f"<@{tech.id}>, this case has been complimented by <@{tech.id}>.",
-            colour=discord.Color.green(),
-            timestamp=now
-        )
-
-        fb_embed.add_field(name="Description", value=str(pp.description), inline=False)
-
-        fb_embed.set_author(name=f"{claim.case_num}", icon_url=f'{tech.display_avatar}')
-
-        # Create thread
-        thread = await case_channel.create_thread(
-            name=f"{claim.case_num}",
-            message=None,
-            auto_archive_duration=4320,
-            type=discord.ChannelType.private_thread,
-            reason="Case has been complimented 😁",
-            invitable=False
-        )
-
-        # Add user to thread and send message
-        await thread.add_user(tech)
-        message = await thread.send(embed=fb_embed, view=KudosView(self))
-
-        # Add a Ping class to store the kudos comment data
-        kudo = Ping(thread.id, message.id, "Kudos", str(pp.description))
-        kudo.add_to_database(self.connection)
-
-        claim.add_ping_thread(self.connection, kudo.thread_id)
-
-        # Remove PendingPing
-        pp.remove_from_database(self.connection)
-
 
     @tasks.loop(seconds=5)  # repeat after every 5 seconds
     async def resend_outages_loop(self):
