+++ bot/forms/kudos_form.py
@@ -11,7 +11,7 @@ from bot.models.pending_ping import PendingPing
 from bot.status import Status
 from bot.views.kudos_view import KudosView
 
-from bot.helpers import is_working_time
+from bot.helpers.other import is_working_time
 
 # Use TYPE_CHECKING to avoid circular import from bot
 from typing import TYPE_CHECKING
